requirements.txt
streamlit
pandas
requests
tensorflow
scikit-learn
numpy
matplotlib
seaborn
google-generativeai
PyPDF2
pillow